+++ src/entities/auth/api/auth.queries.ts
@@ -7,7 +7,6 @@ import { AuthUtil } from '@/shared/utils/auth.util';
 import { ROUTES_PATHS, isProtectedPath } from '@/shared/config/route-paths';
 import { TEXTS } from '@/shared/config/texts';
 import { toast } from '@/shared/lib/toast/toast';
-import { API_ENDPOINTS } from '@/shared/config/api';
 import { SERVER_ERROR_CODE } from '@/shared/config/error-code';
 import { useNavigate } from 'react-router-dom';
 import { requestAndRegisterFcmToken, unregisterFcmToken } from '@/shared/lib/firebase/fcm';
@@ -97,7 +96,7 @@ export const useCreateAccountMutation = () => {
       }
     },
     onSuccess: () => {
-      navigate(API_ENDPOINTS.auth.login);
+      navigate(ROUTES_PATHS.AUTH.LOGIN);
     },
   });
 };